[fix] get/pingpong/match/opponent
