+++ src/main/java/io/pp/arcade/v1/domain/opponent/OpponentRepository.java
@@ -1,5 +1,6 @@
-package io.pp.arcade.v1.domain.slot;
+package io.pp.arcade.v1.domain.opponent;
 
+import io.pp.arcade.v1.domain.opponent.Opponent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
